Hide URL fragment and replace it with Share button
URL contains a seed which must be kept secret. After the page is loaded,
clear all content following the #fragment.

In order to share the URL with others a new button has been added. The
button opens a share card containing wallet's URL. For easier sharing
across devices a QR code is available.